Add new Icon and Animation to Home Screen
